The search controller can list all meals in the database in the table view.
Bugs: In the Meal object, the ingredients are an Array<String>. Since the database gives back the ingredients as a string, I need to convert the string, to an array of strings.
Also, the database code should be refactored so there isn't as much duplicated code. I'm not going to do it yet.